renamed export to clone
