+++ apps/Standalone/src/clonetostandard/app/cloneshell.tsx
@@ -3,18 +3,18 @@ import { ReactQueryProvider } from '@microsoft/logic-apps-designer';
 import { useQuery } from '@tanstack/react-query';
 import { DevToolbox } from '../components/devtoolbox';
 import { loadToken } from '../../environments/environment';
-import { ExportConsumption } from './exportconsumption';
+import { CloneToStandard } from './clonetostandard';
 const LoadWhenArmTokenIsLoaded = ({ children }: { children: ReactNode }) => {
   const { isLoading } = useQuery(['armToken'], loadToken);
   return isLoading ? null : <>{children}</>;
 };
-export const ExportConsumptionWrapper = () => {
+export const CloneToStandardWrapper = () => {
   return (
     <ReactQueryProvider>
       <LoadWhenArmTokenIsLoaded>
         <DevToolbox />
         <div style={{ height: '100vh' }}>
-          <ExportConsumption />
+          <CloneToStandard />
         </div>
       </LoadWhenArmTokenIsLoaded>
     </ReactQueryProvider>
