+++ libs/designer/src/lib/core/clonetostandard/clonedataprovider.tsx
@@ -1,8 +1,8 @@
 import type React from 'react';
 import { useEffect } from 'react';
 import { useDispatch } from 'react-redux';
-import type { AppDispatch } from '../state/exportconsumption/store';
-import { type ResourceState, setResourceData } from '../state/exportconsumption/resourceslice';
+import type { AppDispatch } from '../state/clonetostandard/store';
+import { type ResourceState, setResourceData } from '../state/clonetostandard/resourceslice';
 
 export interface ExportDataProviderProps {
   resourceDetails: ResourceState;
@@ -13,7 +13,7 @@ const DataProviderInner = ({ children }: { children?: React.ReactNode }) => {
   return <>{children}</>;
 };
 
-export const ExportDataProvider = ({ resourceDetails, children }: ExportDataProviderProps) => {
+export const CloneDataProvider = ({ resourceDetails, children }: ExportDataProviderProps) => {
   const dispatch = useDispatch<AppDispatch>();
 
   useEffect(() => {
