+++ libs/designer/src/lib/core/state/clonetostandard/resourceslice.ts
