+++ libs/designer/src/lib/ui/index.tsx
@@ -4,7 +4,7 @@ export * from './panel';
 export * from './menuItems';
 export * from './templates';
 export * from './mcp/wizard/McpWizard';
-export * from './exportconsumption/wizard/exportwizard';
+export * from './clonetostandard/wizard/clonewizard';
 export * from './dialog/combineVariableDialog';
 export * from './dialog/triggerDescriptionDialog';
 export { getLogicAppsCategories } from './configuretemplate/resources';
